divide the redux actions in two files
