+++ src/components/UserLogin/CreateUserForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { createUser } from '../../redux/postActions';
+import { createUser } from '../../redux/authActions';
 import { useNavigate } from 'react-router-dom';
 import './Login.css'
